+++ docs/changelog.md
@@ -5,6 +5,7 @@
 ### New
 
 * Initial support for the sRGB color space:
+  - added `ColorCalc` exception interface (for type-hinting)
   - added `ColorFormat` constant container
   - added `CssColorFunctions` map
   - added `CssColorKeywords` map
@@ -12,4 +13,6 @@
   - added `IsCssColorFunction` check
   - added `IsCssColorKeyword` check
   - added `IsCssColorValue` check
+  - added `MapCssColorKeywordToValue`
   - added `RemoveHashPrefix` value editor
+  - added `UnknownCssColorKeyword` exception